+++ core/src/Stripes.cpp
@@ -8,7 +8,7 @@ namespace core {
     /* POTENTIALS */
 
     template <typename T>
-    uint8_t Stripes<T>::computeBits(T act, T wgt, uint8_t act_prec, uint8_t wgt_prec, uint8_t network_bits) {
+    uint16_t Stripes<T>::computeBits(T act, T wgt, uint8_t act_prec, uint8_t wgt_prec, uint8_t network_bits) {
         return act_prec * network_bits;
     }
